Update Steamworks to v1.48a
We also switch over to the new Steamworks provided interface getters, which allows removing some C++ glue code.

+++ src/steam/remote_storage.rs
@@ -23,7 +23,7 @@ impl UgcHandle {
                     sys::RemoteStorageDownloadUGCResult_t_k_iCallback,
                     || {
                         sys::SteamAPI_ISteamRemoteStorage_UGCDownloadToLocation(
-                            client.0.remote_storage as isize,
+                            client.0.remote_storage,
                             self.0,
                             location.as_ptr(),
                             priority,
